feat(asset): 增加 asset controller，调整日志记录方式

+++ internal/apiserver/service/types.go
@@ -1,6 +1,6 @@
 package service
 
 type ListOptions struct {
-	Limit  int64
-	Offset int64
+	Limit  int
+	Offset int
 }
